+++ emulator/crates/ir_macros/src/impl/r.rs
@@ -17,9 +17,9 @@ pub fn proc_macro_impl(args: TokenStream, ast: ItemStruct) -> TokenStream {
         #vis struct #ident {
             opcode: u8,
             opcode_sub: u8,
-            rd: u8,
-            rs1: u8,
-            rs2: u8,
+            rd: usize,
+            rs1: usize,
+            rs2: usize,
         }
 
         impl From<u64> for #ident {
@@ -27,9 +27,9 @@ pub fn proc_macro_impl(args: TokenStream, ast: ItemStruct) -> TokenStream {
                 Self {
                     opcode: #opcode,
                     opcode_sub: #opcode_sb,
-                    rd:         ((raw >>  8) &    0b11111) as u8,
-                    rs1:        ((raw >> 13) &    0b11111) as u8,
-                    rs2:        ((raw >> 18) &    0b11111) as u8,
+                    rd:         ((raw >>  8) &    0b11111) as usize,
+                    rs1:        ((raw >> 13) &    0b11111) as usize,
+                    rs2:        ((raw >> 18) &    0b11111) as usize,
                 }
             }
         }
